+++ tests/unit/resources/networking/test_logical_interconnects.py
@@ -375,3 +375,16 @@ class LogicalInterconnectsTest(unittest.TestCase):
 
         expected_uri = '{}/igmpSettings'.format(self.uri)
         mock_get.assert_called_once_with(expected_uri)
+
+    @mock.patch.object(ResourceHelper, 'do_post')
+    def test_bulk_inconsistency_validate(self, mock_post):
+        resource = {
+            "logicalInterconnectUris": [
+                "/rest/logical-interconnects/d0432852-28a7-4060-ba49-57ca973ef6c2"
+        ]
+        }
+        resource = resource.copy()
+
+        self._logical_interconnect.bulk_inconsistency_validate(resource)
+
+        mock_post.assert_called_once_with(resource, uri='{}/bulk-inconsistency-validation'.format(self.uri), force=False, timeout=-1)